% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.factor.data.frame}
\alias{as.factor.data.frame}
\title{Convert all columns of a dataframe into factors}
\usage{
as.factor.data.frame(df, levels, ...)
}
\arguments{
\item{df}{A dataframe}

\item{levels}{A character vector of factor levels}

\item{...}{Arguments that can be passed to factor()}
}
\value{
A factor dataframe
}
\description{
Convert all columns of a dataframe into factors
}
