% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cacheAndReadData}
\alias{cacheAndReadData}
\title{Cache remote files and read as a dataframe}
\usage{
cacheAndReadData(remote.path, local.path = NULL, overwrite = F)
}
\arguments{
\item{remote.path}{Remote path to txt or rds file}

\item{local.path}{Local path to copy to}

\item{overwrite}{Redownload data from remote?}
}
\value{
A dataframe
}
\description{
Cache remote files and read as a dataframe
}
