% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChromArmPloidies.R
\name{calcChromArmCnChange}
\alias{calcChromArmCnChange}
\title{Calculate change in chrom arm copy number compared to genome copy number}
\usage{
calcChromArmCnChange(x)
}
\arguments{
\item{x}{Output from calcChromArmPloidies() as a matrix or vector}
}
\value{
A matrix or vector
}
\description{
Calculate change in chrom arm copy number compared to genome copy number
}
