% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.effSize.R
\name{cliffDelta}
\alias{cliffDelta}
\alias{cliffDelta.default}
\alias{cliffDelta.matrix}
\alias{cliffDelta.data.frame}
\title{Calculate Cliff's delta values (for continuous data)}
\usage{
cliffDelta(x, ...)

\method{cliffDelta}{default}(x, y)

\method{cliffDelta}{matrix}(x, y)

\method{cliffDelta}{data.frame}(x, y)
}
\arguments{
\item{x}{A numeric vector for the 1st group of observations. Alternatively, a matrix, where
comparisons will be performed by row.}

\item{y}{Same as x but for the 2nd group of observations.}
}
\value{
A numeric vector of Cliff's delta values
}
\description{
Calculate effect sizes using Cliff's delta values for pairwise continuous variable
comparisons (e.g. comparisons that would be done by a wilcox test or t-test)
}
