% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFeatures.R
\name{combineFeatures}
\alias{combineFeatures}
\alias{combineFeatures.matrix}
\alias{combineFeatures.data.frame}
\alias{combineFeatures.numeric}
\alias{combineFeatures.integer}
\title{Combine feature columns by summing them together}
\usage{
combineFeatures(x, ...)

\method{combineFeatures}{matrix}(x, target.features = NULL, regex = NULL, target.name = NULL)

\method{combineFeatures}{dataframe}(x, target.features = NULL, regex = NULL, target.name = NULL)

\method{combineFeatures}{numeric}(x, target.features = NULL, regex = NULL, target.name = NULL)

\method{combineFeatures}{integer}(x, target.features = NULL, regex = NULL, target.name = NULL)
}
\arguments{
\item{x}{An integer/numeric vector, matrix or dataframe}

\item{target.features}{A character vector of column (feature) names}

\item{regex}{Instead of target.features, a regex can be specified}

\item{target.name}{A name to assign the new feature. If unspecified, will use the name of the
first old feature}
}
\value{
The original data with the indicated features combined
}
\description{
Combine feature columns by summing them together
}
