% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{effectSize}
\alias{effectSize}
\title{Calculate effect size}
\usage{
effectSize(xpos = NULL, xneg = NULL, ypos = NULL, yneg = NULL, m = NULL)
}
\arguments{
\item{xpos}{Group 'x' responders}

\item{xneg}{Group 'x' non responders}

\item{ypos}{Group 'y' responders}

\item{yneg}{Group 'y' non responders}

\item{m}{A matrix containing the xpos, xneg, ypos, yneg (in this order)}
}
\value{
A numeric vector
}
\description{
Effect size calculation is based on odds ratio = (xpos/xneg) / (ypos/yneg). Odds
ratios < 1 are scaled to become negative values.
}
