% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPlotsAsPngPdf.R
\name{exportPlotsAsPngPdf}
\alias{exportPlotsAsPngPdf}
\title{Export a list of plots as a PDF with a PNG per page}
\usage{
exportPlotsAsPngPdf(
  plot.list,
  out.path,
  page.width = 10,
  page.height = 4,
  res = 300
)
}
\arguments{
\item{plot.list}{A list of plots}

\item{out.path}{Output PDF path}

\item{page.width}{Page width in inches}

\item{page.height}{Page height in inches}

\item{res}{Page resolution}
}
\description{
Export a list of plots as a PDF with a PNG per page
}
