% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.legacy.R
\name{fisherTestApply}
\alias{fisherTestApply}
\title{Perform multiple fisher's exact tests}
\usage{
fisherTestApply(m, verbose = F, ...)
}
\arguments{
\item{m}{A matrix containing the 4 columns corresponding to a contingency matrix:
xpos: Group 'x' responders
xneg: Group 'x' non responders
ypos: Group 'y' responders
yneg: Group 'y' non responders}

\item{verbose}{Show progress?}

\item{...}{Arguments that can be passed to fisher.test()}
}
\value{
A numeric vector of pvalues
}
\description{
Perform multiple fisher's exact tests
}
