% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeUtils.R
\name{getChromArm}
\alias{getChromArm}
\title{Get chromosome arm}
\usage{
getChromArm(
  df = NULL,
  centro.pos = CENTRO_POS_HG19,
  one.armed.chroms = c(13, 14, 15, 21, 22),
  arm.only = F,
  seq.levels.style = "NCBI",
  centro.intervals.rough.fix = F,
  show.warnings = F
)
}
\arguments{
\item{df}{A dataframe containing the columns: chrom, start, end}

\item{centro.pos}{A named integer vector of the centromere positions (default: hg19)}

\item{arm.only}{Only return the arm and not the chrom number?}

\item{centro.intervals.rough.fix}{If TRUE, intervals intersecting with the centromere will be
assigned to the q arm.}

\item{show.warnings}{Show warning messages?}

\item{chrom.arm.names}{A character vector in the form c('1p','1q','2p','2q', ...). The default
'auto' means that the human chromosome arm names are used. Note that chroms 13, 14, 15, 21, 22
are considered to only have the long (i.e. q) arm.}
}
\value{
A character vector of chromosome arm names
}
\description{
Get chromosome arm
}
