% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeUtils.R
\name{getCytoband}
\alias{getCytoband}
\title{Get cytoband}
\usage{
getCytoband(
  df,
  cytobands.path = CYTOBANDS,
  greedy.multi.match = F,
  verbose = F
)
}
\arguments{
\item{df}{A dataframe containing the columns: chrom, start, end}

\item{cytobands.path}{Path to the cytobands txt file with the columns: chrom, start, end, band, stain}

\item{greedy.multi.match}{If TRUE, when multiple cytobands match to a region, only the first
cytoband will be reported}

\item{verbose}{Show progress?}
}
\value{
A character vector
}
\description{
Get cytoband
}
