% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{insColAfter}
\alias{insColAfter}
\title{Insert column(s) after a column of an existing dataframe}
\usage{
insColAfter(df, v, after, colname = NULL)
}
\arguments{
\item{df}{Input dataframe}

\item{v}{A vector or dataframe}

\item{after}{Column index or name}

\item{colname}{Set inserted column name. Has no effect if v is not a vector}
}
\value{
A dataframe with the inserted column(s)
}
\description{
Insert column(s) after a column of an existing dataframe
}
