% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{insertRow}
\alias{insertRow}
\title{Insert a row into a dataframe}
\usage{
insertRow(df, row.num, new.row = NULL, offset = 1)
}
\arguments{
\item{df}{A dataframe}

\item{row.num}{Row index to insert the row. Rows below this are shifted down.}

\item{new.row}{The row as a vector or dataframe to insert}

\item{offset}{Insert the row at +offset rows after row.num}
}
\value{
A dataframe
}
\description{
Insert a row into a dataframe
}
