% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeUtils.R
\name{isOverlapping}
\alias{isOverlapping}
\title{Find overlapping intervals}
\usage{
isOverlapping(start1, end1, start2, end2, verbose = F)
}
\arguments{
\item{start1}{A vector of start positions (#1)}

\item{end1}{A vector of end positions (#1)}

\item{start2}{A vector of start positions (#2)}

\item{end2}{A vector of end positions (#2)}

\item{verbose}{Show progress?}
}
\value{
A list of indexes of #1 for which #2 overlaps
}
\description{
Find overlapping intervals
}
