% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeUtils.R
\name{isOverlappingChromPos}
\alias{isOverlappingChromPos}
\title{Find overlapping genomic intervals}
\usage{
isOverlappingChromPos(
  chrom1 = NULL,
  start1 = NULL,
  end1 = NULL,
  chrom2 = NULL,
  start2 = NULL,
  end2 = NULL,
  df1 = NULL,
  df2 = NULL
)
}
\arguments{
\item{chrom1}{A vector or chromosome names (#1)}

\item{start1}{A vector of start positions (#1)}

\item{end1}{A vector of end positions (#1)}

\item{chrom2}{A vector or chromosome names (#1)}

\item{start2}{A vector of start positions (#2)}

\item{end2}{A vector of end positions (#2)}
}
\value{
A list of indexes of #1 for which #2 overlaps
}
\description{
Find overlapping genomic intervals
}
