% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.effSize.R
\name{oddsRatio}
\alias{oddsRatio}
\alias{oddsRatio.default}
\alias{oddsRatio.matrix}
\alias{oddsRatio.data.frame}
\title{Calculate odds ratio}
\usage{
oddsRatio(x, ...)

\method{oddsRatio}{default}(xpos, xneg, ypos, yneg, trans.method = "none", logistic.growth = 0.2)

\method{oddsRatio}{matrix}(m, ...)

\method{oddsRatio}{dataframe}(m, ...)
}
\arguments{
\item{xpos}{Group 'x' responders}

\item{xneg}{Group 'x' non responders}

\item{ypos}{Group 'y' responders}

\item{yneg}{Group 'y' non responders}

\item{trans.method}{Can be 'none', 'linear' or 'logistic'}

\item{logistic.growth}{Steepness/gradient of the logistic function when trans.method=='logistic'}
}
\value{
A numeric vector
}
\description{
Effect size calculation is based on odds ratio = (xpos/xneg) / (ypos/yneg). When
trans.method=='linear' odds ratios < 1 are scaled to become negative values. When
trans.method=='logistic' the linearly scaled odds ratios are logistically transformed with the
equation 2 / (1 + exp(1)^(-logistic.growth*OR) ) - 1
}
