% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRmdProfile.R
\name{plotRmdProfile}
\alias{plotRmdProfile}
\title{Plot regional mutational density profiles}
\usage{
plotRmdProfile(
  m,
  top.n = 15,
  plot.title = "RMD profile",
  y.lab = "Probability"
)
}
\arguments{
\item{m}{A matrix where rows are samples and columns are RMD bins}

\item{top.n}{Show labels for the features with the highest peaks}

\item{plot.title}{Plot title}

\item{y.lab}{y axis title}
}
\value{
A ggplot object
}
\description{
Plot regional mutational density profiles
}
