% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rle2Clusters}
\alias{rle2Clusters}
\title{Assign clusters to a vector of values based on run length encoding}
\usage{
rle2Clusters(x)
}
\arguments{
\item{x}{An integer vector}
}
\value{

}
\description{
Assign clusters to a vector of values based on run length encoding
}
