% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRequiredCols.R
\name{selectRequiredCols}
\alias{selectRequiredCols}
\title{Check column names of an input table}
\usage{
selectRequiredCols(df, required.cols, sel.cols = NULL)
}
\arguments{
\item{df}{An input dataframe}

\item{required.cols}{A character vector indicating the require col names}

\item{sel.cols}{A named character vector with the names. Each name corresponds to the names of
the `required.cols`. Each value corresponds to the the column names in the txt file. If
`sel.cols` is NULL, this will automatically be generated from required.cols}
}
\value{
A subset dataframe if all checks pass
}
\description{
This function checks whether all `names(sel.cols)` are in `required.cols`. If this
passes, the function will check whether all `sel.cols` values are in `colnames(df)`. If this,
passes, a dataframe will be returned with the selected columns
}
