% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFeatures.R
\name{splitFeaturesByGroup}
\alias{splitFeaturesByGroup}
\alias{splitFeaturesByGroup.default}
\alias{splitFeaturesByGroup.data.frame}
\title{Split a dataframe into feature groups}
\usage{
splitFeaturesByGroup(x, ...)

\method{splitFeaturesByGroup}{default}(x, rm.tags = F)

\method{splitFeaturesByGroup}{data.frame}(x, rm.tags = F)
}
\arguments{
\item{rm.tags}{Remove feature group tags?}

\item{df}{A dataframe where colnames are in the form
'feature_tag.feature_name'}
}
\value{
A list of dataframes
}
\description{
Split a dataframe into feature groups
}
