% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{wilcoxTestApply}
\alias{wilcoxTestApply}
\title{Perform multiple wilcox tests}
\usage{
wilcoxTestApply(l, verbose = F, ...)
}
\arguments{
\item{l}{A list with pairs of numeric vectors:
list(
  first=list(numeric(), numeric()),
  second=list(numeric(), numeric()),
  third=list(numeric(), numeric())
)}

\item{verbose}{Show progress?}

\item{...}{}
}
\value{
A vector of pvalues
}
\description{
Perform multiple wilcox tests
}
