% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write.tsv}
\alias{write.tsv}
\title{Write tsv file}
\usage{
write.tsv(x, file, ...)
}
\arguments{
\item{x}{The object to be written, preferably a matrix or data frame. If not, it is attempted to
coerce x to a data frame.}

\item{file}{Path to the output file. If file ends with .gz, a gzip compressed file will be written}

\item{...}{Arguments that can be passed to write.table()}
}
\value{

}
\description{
Write tsv file
}
